requirements.txt
flask
pillow
pytesseract
pymupdf
reportlab